# [WIP] Vectorize EncodeAsciiCharsToBytes