--- a/src/Kestrel.Core/Internal/Http/PipelineExtensions.cs
+++ b/src/Kestrel.Core/Internal/Http/PipelineExtensions.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Buffers;
 using System.IO.Pipelines;
+using System.Numerics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -182,87 +183,117 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
             const int Shift16Shift24 = (1 << 16) | (1 << 24);
             const int Shift8Identity = (1 << 8) | (1);
 
-            // Encode as bytes upto the first non-ASCII byte and return count encoded
-            int i = 0;
-            // Use Intrinsic switch
-            if (IntPtr.Size == 8) // 64 bit
+            // Calcuate end position
+            var end = input + length;
+            do
             {
-                if (length < 4) goto trailing;
+                if (!Vector.IsHardwareAccelerated || input > end - Vector<ushort>.Count * 2)
+                {
+                    length = (int)(end - input);
+                    int i = 0;
+                    // Use Intrinsic switch
+                    if (IntPtr.Size == 8) // 64 bit
+                    {
+                        if (length < 4) goto trailing;
 
-                int unaligned = (int)(((ulong)input) & 0x7) >> 1;
-                // Unaligned chars
-                for (; i < unaligned; i++)
-                {
-                    char ch = *(input + i);
-                    *(output + i) = (byte)ch; // Cast convert
+                        int unaligned = (int)(((ulong)input) & 0x7) >> 1;
+                        // Unaligned chars
+                        for (; i < unaligned; i++)
+                        {
+                            char ch = input[i];
+                            output[i] = (byte)ch; // Cast convert
+                        }
+
+                        // Aligned
+                        int ulongDoubleCount = (length - i) & ~0x7;
+                        for (; i < ulongDoubleCount; i += 8)
+                        {
+                            ulong inputUlong0 = *(ulong*)(input + i);
+                            ulong inputUlong1 = *(ulong*)(input + i + 4);
+                            // Pack 16 ASCII chars into 16 bytes
+                            *(uint*)(output + i) =
+                                ((uint)((inputUlong0 * Shift16Shift24) >> 24) & 0xffff) |
+                                ((uint)((inputUlong0 * Shift8Identity) >> 24) & 0xffff0000);
+                            *(uint*)(output + i + 4) =
+                                ((uint)((inputUlong1 * Shift16Shift24) >> 24) & 0xffff) |
+                                ((uint)((inputUlong1 * Shift8Identity) >> 24) & 0xffff0000);
+                        }
+                        if (length - 4 > i)
+                        {
+                            ulong inputUlong = *(ulong*)(input + i);
+                            // Pack 8 ASCII chars into 8 bytes
+                            *(uint*)(output + i) =
+                                ((uint)((inputUlong * Shift16Shift24) >> 24) & 0xffff) |
+                                ((uint)((inputUlong * Shift8Identity) >> 24) & 0xffff0000);
+                            i += 4;
+                        }
+
+                        trailing:
+                        for (; i < length; i++)
+                        {
+                            char ch = input[i];
+                            output[i] = (byte)ch; // Cast convert
+                        }
+
+                        return;
+                    }
+                    else // 32 bit
+                    {
+                        // Unaligned chars
+                        if ((unchecked((int)input) & 0x2) != 0)
+                        {
+                            char ch = input[0];
+                            i = 1;
+                            output[0] = (byte)ch; // Cast convert
+                        }
+
+                        // Aligned
+                        int uintCount = (length - i) & ~0x3;
+                        for (; i < uintCount; i += 4)
+                        {
+                            uint inputUint0 = *(uint*)(input + i);
+                            uint inputUint1 = *(uint*)(input + i + 2);
+                            // Pack 4 ASCII chars into 4 bytes
+                            *(ushort*)(output + i) = (ushort)(inputUint0 | (inputUint0 >> 8));
+                            *(ushort*)(output + i + 2) = (ushort)(inputUint1 | (inputUint1 >> 8));
+                        }
+                        if (length - 1 > i)
+                        {
+                            uint inputUint = *(uint*)(input + i);
+                            // Pack 2 ASCII chars into 2 bytes
+                            *(ushort*)(output + i) = (ushort)(inputUint | (inputUint >> 8));
+                            i += 2;
+                        }
+
+                        if (i < length)
+                        {
+                            char ch = input[i];
+                            output[i] = (byte)ch; // Cast convert
+                            i = length;
+                        }
+                    }
+
+                    return;
                 }
 
-                // Aligned
-                int ulongDoubleCount = (length - i) & ~0x7;
-                for (; i < ulongDoubleCount; i += 8)
+                // do/while as entry condition already checked
+                do
                 {
-                    ulong inputUlong0 = *(ulong*)(input + i);
-                    ulong inputUlong1 = *(ulong*)(input + i + 4);
-                    // Pack 16 ASCII chars into 16 bytes
-                    *(uint*)(output + i) =
-                        ((uint)((inputUlong0 * Shift16Shift24) >> 24) & 0xffff) |
-                        ((uint)((inputUlong0 * Shift8Identity) >> 24) & 0xffff0000);
-                    *(uint*)(output + i + 4) =
-                        ((uint)((inputUlong1 * Shift16Shift24) >> 24) & 0xffff) |
-                        ((uint)((inputUlong1 * Shift8Identity) >> 24) & 0xffff0000);
-                }
-                if (length - 4 > i)
-                {
-                    ulong inputUlong = *(ulong*)(input + i);
-                    // Pack 8 ASCII chars into 8 bytes
-                    *(uint*)(output + i) =
-                        ((uint)((inputUlong * Shift16Shift24) >> 24) & 0xffff) |
-                        ((uint)((inputUlong * Shift8Identity) >> 24) & 0xffff0000);
-                    i += 4;
-                }
+                    // Convert 2 char vectors to 1 byte Vector at a time
+                    var bytes = Vector.Narrow(
+                        Unsafe.AsRef<Vector<ushort>>(input), 
+                        Unsafe.AsRef<Vector<ushort>>(input + Vector<ushort>.Count));
+                    Unsafe.AsRef<Vector<byte>>(output) = bytes;
 
-                trailing:
-                for (; i < length; i++)
-                {
-                    char ch = *(input + i);
-                    *(output + i) = (byte)ch; // Cast convert
-                }
-            }
-            else // 32 bit
-            {
-                // Unaligned chars
-                if ((unchecked((int)input) & 0x2) != 0)
-                {
-                    char ch = *input;
-                    i = 1;
-                    *(output) = (byte)ch; // Cast convert
-                }
+                    input += Vector<ushort>.Count * 2;
+                    output += Vector<byte>.Count;
 
-                // Aligned
-                int uintCount = (length - i) & ~0x3;
-                for (; i < uintCount; i += 4)
-                {
-                    uint inputUint0 = *(uint*)(input + i);
-                    uint inputUint1 = *(uint*)(input + i + 2);
-                    // Pack 4 ASCII chars into 4 bytes
-                    *(ushort*)(output + i) = (ushort)(inputUint0 | (inputUint0 >> 8));
-                    *(ushort*)(output + i + 2) = (ushort)(inputUint1 | (inputUint1 >> 8));
-                }
-                if (length - 1 > i)
-                {
-                    uint inputUint = *(uint*)(input + i);
-                    // Pack 2 ASCII chars into 2 bytes
-                    *(ushort*)(output + i) = (ushort)(inputUint | (inputUint >> 8));
-                    i += 2;
-                }
+                    // Loop while remaining double Vector sizes
+                } while (input <= end - Vector<ushort>.Count * 2);
 
-                if (i < length)
-                {
-                    char ch = *(input + i);
-                    *(output + i) = (byte)ch; // Cast convert
-                    i = length;
-                }
-            }
+                // Vector path done, loop back to do non-Vector
+                // If is a exact multiple of vector size, bail now
+            } while (input < end);
         }
 
         private static byte[] NumericBytesScratch => _numericBytesScratch ?? CreateNumericBytesScratch();