Use `Vector.Narrow` since its in 2.1

Ignoring whitespace is easier on the eye https://github.com/aspnet/KestrelHttpServer/pull/2339/files?w=1