@benaadams 2018-02-21
ResponseHeadersWritingBenchmark

Before
```
 Method |                       Type |       Mean |     Error |   StdDev |        Op/s | Allocated |
------- |--------------------------- |-----------:|----------:|---------:|------------:|----------:|
 Output |                 LiveAspNet | 1,958.3 ns |  8.330 ns | 12.47 ns |   510,642.5 |     363 B |
 Output |           PlaintextChunked | 1,274.2 ns |  7.349 ns | 11.00 ns |   784,804.5 |     202 B |
 Output | PlaintextChunkedWithCookie | 1,687.5 ns |  7.947 ns | 11.89 ns |   592,604.8 |     497 B |
 Output |        PlaintextWithCookie | 1,358.4 ns | 12.269 ns | 18.36 ns |   736,135.5 |     437 B |
 Output |       TechEmpowerPlaintext |   919.0 ns | 11.073 ns | 16.57 ns | 1,088,097.9 |     143 B |
```
After
```
 Method |                       Type |       Mean |     Error |    StdDev |        Op/s | Allocated |
------- |--------------------------- |-----------:|----------:|----------:|------------:|----------:|
 Output |                 LiveAspNet | 1,964.1 ns |  9.683 ns | 14.493 ns |   509,137.7 |     363 B |
 Output |           PlaintextChunked | 1,257.3 ns |  5.396 ns |  8.076 ns |   795,328.3 |     202 B |
 Output | PlaintextChunkedWithCookie | 1,648.0 ns | 22.138 ns | 33.135 ns |   606,803.1 |     497 B |
 Output |        PlaintextWithCookie | 1,256.3 ns | 16.669 ns | 24.950 ns |   795,971.1 |     437 B |
 Output |       TechEmpowerPlaintext |   908.5 ns |  6.936 ns | 10.382 ns | 1,100,667.5 |     143 B |
```

Hmm... I made it too good before, marking this WIP